Consolidate jobseeker support
