+++ openfisca_aotearoa/tests/social_security/jobseeker_support/abatement.yaml
@@ -1,0 +1,59 @@
+- name: >
+    Social Security Act 2018 -
+    Schedule 2 Dictionary -
+    Income test 1 -
+    The applicable rate of benefit must be reduced—
+      (a) by 30 cents for every $1 of the total income of the beneficiary and
+          the beneficiary’s spouse or partner that is more than $160 a week but
+          not more than $250 a week
+      (b) by 70 cents for every $1 of that income that is more than $250 a week
+  period: 2018-11-26
+  input:
+    jobseeker_support__income:
+      2023-W01: [159.99, 160.00, 160.01, 161.99, 249.99, 250.00, 250.01, 251.99]
+    schedule_4__part1_1_c:
+      2023-W01: [true, true, true, true, true, true, true, true]
+  output:
+    jobseeker_support__abatement:
+      2023-W01: [0, 0, 0, 0.30, 26.70, 27.00, 27.00, 27.70]
+
+
+- name: >
+    Social Security Act 2018 -
+    Schedule 2 Dictionary -
+    Income test 3 -
+    The applicable rate of benefit must be reduced by 70 cents for every $1 of
+    the total income of the beneficiary and the beneficiary’s spouse or partner
+    that is more than,—
+      (b) [...] $160 a week
+  period: 2018-11-26
+  input:
+    jobseeker_support__income:
+      2023-W01: [159.99, 160.00, 160.01, 161.99, 249.99, 250.00, 250.01, 251.99]
+    schedule_4__part1_1_a:
+      2023-W01: [true, true, true, true, true, true, true, true]
+  output:
+    jobseeker_support__abatement:
+      2023-W01: [0, 0, 0, 0.70, 62.30, 63.00, 63.00, 63.70]
+
+
+- name: >
+    Social Security Act 2018 -
+    Schedule 2 Dictionary -
+    Income test 4 -
+    The applicable rate of benefit must be reduced by 35 cents for every $1 of
+    the total income of the beneficiary and the beneficiary’s spouse or partner
+    that is more than $160 a week
+  period: 2018-11-26
+  input:
+    jobseeker_support__income:
+      2023-W01: [159.99, 160.00, 160.01, 161.99, 249.99, 250.00, 250.01, 251.99]
+    schedule_4__part1_1_d:
+      2023-W01: [false, false, false, false, false, false, false, false]
+    schedule_4__part1_1_g:
+      2023-W01: [true, true, true, true, true, true, true, true]
+  output:
+    jobseeker_support__abatement:
+      2023-W01: [0, 0, 0, 0.35, 31.15, 31.50, 31.50, 31.85]
+    schedule_4__part1_1_d:
+      2023-W01: [false, false, false, false, false, false, false, false]
