+++ openfisca_aotearoa/tests/social_security/jobseeker_support/base.yaml
@@ -216,13 +216,6 @@
       2022-W01: [240.65, 406.78, 239.70]
       2023-W01: [283.00, 440.96, 274.37]
       2023-W15: [303.43, 472.79, 294.18]
-    jobseeker_support__abatement:
-      2020-W01: [143.5, 0, 0]
-      2021-W01: [59.5, 0, 0]
-      2021-W23: [0, 0, 0]
-      2022-W01: [138.6, 0, 0]
-      2023-W01: [31.5, 0, 0]
-      2023-W15: [31.5, 0, 0]
     jobseeker_support__benefit:
       2020-W01: [52, 0, 0]
       2021-W01: [154.53, 0, 0]
